+++ lib/glimmer/libui/tab_item_proxy.rb
@@ -54,6 +54,7 @@ module Glimmer
       end
       alias set_margined margined
       alias margined= margined
+      alias margined? margined
           
       private